requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
docker==4.4.4
Flask==2.0.1
Flask-JWT==0.3.2
flask-restx==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
PyJWT==1.4.2
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
pytest-docker-tools==3.1.0
pytest-flask==1.2.0
pytest-flask-sqlalchemy==1.0.2
pytest-mock==3.6.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
SQLAlchemy-Utils==0.37.8
toml==0.10.2
urllib3==1.26.6
websocket-client==1.1.1
Werkzeug==2.0.1
